refactor: rename components for readbility
