+++ packages/react/src/components/accordion/accordion.tsx
@@ -1,68 +1,97 @@
-import React from 'react';
-import styled from 'styled-components';
-import { AccordionProps } from './accordion-types';
-import { AccordionSectionStyled, HeadingStyled, ButtonStyled } from './accordion-styles';
-import { Button } from '../buttons/button';
-import { Icon } from '../icon/icon';
-import { Heading } from '../heading/heading';
-
-const AccordionSection = styled.section<{ isExpanded: boolean }>`
-    ${({ isExpanded, theme }) => AccordionSectionStyled(isExpanded, theme)};
-`;
-
-const StyledHeading = styled(Heading)`
-    ${HeadingStyled};
-`;
-
-const StyledButton = styled(Button)<{ isExpanded: boolean }>`
-    ${({ isExpanded, theme }) => ButtonStyled(isExpanded, theme)};
-`;
+import React, { useCallback, useState, ReactElement, createRef, useMemo } from 'react';
+import { AccordionItemProps, AccordionProps } from './accordion-types';
+import { AccordionItem } from './accordion-item';
+import { StyledAccordionGroup } from './accordion-styles';
 
 export const Accordion: React.FC<AccordionProps> = ({
-    title,
-    id,
-    type = 'medium',
-    tag = 'h3',
-    isExpanded = false,
-    onToggle = () => {},
-    disabled,
-    children,
-    onKeyDown,
-    buttonRef,
+    id, mode = 'single', children,
 }) => {
-    const headerId = id;
-    const panelId = `panel-${id}`;
+    const generateId = useCallback(
+        (childProps: AccordionItemProps, index: number): string => childProps.id || `${id}-${index}`,
+        [id],
+    );
+
+    const isAccordion = (child: React.ReactNode): child is ReactElement<AccordionItemProps> => (
+        React.isValidElement<AccordionItemProps>(child) && child.type === AccordionItem);
+
+    const filteredChildren = React.Children.toArray(children).filter(isAccordion);
+
+    const [expandedItemIds, setExpandedItemIds] = useState<string[]>(() => (
+        filteredChildren
+            .map((child, index) => {
+                const childProps = child.props;
+                return childProps.isExpanded ? generateId(childProps, index) : null;
+            })
+            .filter((expandedId) => expandedId !== null) as string[]
+    ));
+
+    const handleToggle = useCallback(
+        (itemId: string): void => {
+            if (mode === 'single') {
+                setExpandedItemIds((prevIds) => (prevIds.includes(itemId) ? [] : [itemId]));
+            } else if (mode === 'multi') {
+                setExpandedItemIds((prevIds) => {
+                    if (prevIds.includes(itemId)) {
+                        return prevIds.filter((expandedId) => expandedId !== itemId);
+                    }
+                    return [...prevIds, itemId];
+                });
+            }
+        },
+        [mode, setExpandedItemIds],
+    );
+
+    const childrenArray: ReactElement<AccordionItemProps>[] = useMemo(
+        () => filteredChildren.map((child, index) => {
+            const buttonRef = createRef<HTMLButtonElement>();
+            const childProps = child.props;
+            const accordionId = generateId(childProps, index);
+            const accordionProps: AccordionItemProps = {
+                ...childProps,
+                buttonRef,
+                id: accordionId,
+                onToggle: () => handleToggle(accordionId),
+            };
+            return React.cloneElement(child, accordionProps);
+        }),
+        [filteredChildren, generateId, handleToggle],
+    );
+
+    const handleButtonKeyDown = (
+        event: React.KeyboardEvent<HTMLButtonElement>,
+        index: number,
+    ): void => {
+        const { key } = event;
+        if (key === 'ArrowUp' || key === 'ArrowDown') {
+            event.preventDefault();
+            let newIndex;
+            if (key === 'ArrowUp') {
+                newIndex = index - 1;
+                if (newIndex < 0) {
+                    newIndex = childrenArray.length - 1;
+                }
+            } else {
+                newIndex = (index + 1) % childrenArray.length;
+            }
+            childrenArray[newIndex]?.props?.buttonRef?.current?.focus();
+        }
+    };
 
     return (
-        <>
-            <StyledHeading className="accordion-header" type={type} tag={tag} noMargin>
-                <StyledButton
-                    id={headerId}
-                    className="accordion-button"
-                    buttonType="tertiary"
-                    label={title}
-                    aria-expanded={isExpanded}
-                    aria-controls={panelId}
-                    onClick={() => onToggle()}
-                    isExpanded={isExpanded}
-                    disabled={disabled}
-                    onKeyDown={onKeyDown}
-                    ref={buttonRef}
-                >
-                    <Icon name={isExpanded ? 'caretDown' : 'caretRight'} aria-hidden="true" />
-                </StyledButton>
-            </StyledHeading>
-            <AccordionSection
-                className="accordion-content"
-                id={panelId}
-                aria-labelledby={headerId}
-                aria-expanded={isExpanded}
-                role="region"
-                isExpanded={isExpanded}
-                aria-disabled={disabled}
-            >
-                {children}
-            </AccordionSection>
-        </>
+        <StyledAccordionGroup className='accordion'>
+            {childrenArray.map((accordion, index) => {
+                const { id: accordionId, ...restProps } = accordion.props;
+                const isExpanded = accordionId ? expandedItemIds.includes(accordionId) : false;
+                const accordionItemProps: AccordionItemProps = {
+                    ...restProps,
+                    onKeyDown: (event: React.KeyboardEvent<HTMLButtonElement>) => handleButtonKeyDown(event, index),
+                    isExpanded,
+                    onToggle: accordion.props.onToggle,
+                    buttonRef: accordion.props.buttonRef,
+                    disabled: accordion.props.disabled,
+                };
+                return React.cloneElement(accordion, accordionItemProps);
+            })}
+        </StyledAccordionGroup>
     );
 };
