+++ packages/react/src/components/accordion/index.tsx
@@ -1,2 +1,2 @@
+export { AccordionItem } from './accordion-item';
 export { Accordion } from './accordion';
-export { AccordionContainer } from './accordion-container';
