+++ packages/react/src/index.ts
@@ -40,7 +40,7 @@ export { IntlProvider } from './components/internationalization-provider/interna
 export { Listbox, ListboxOption } from './components/listbox/listbox';
 
 // Miscellaneous
-export { Accordion, AccordionContainer } from './components/accordion/index';
+export { AccordionItem, Accordion } from './components/accordion/index';
 export { Badge } from './components/badge/badge';
 export { Card } from './components/card/card';
 export { CardLink } from './components/card-link/card-link';
