+++ packages/storybook/stories/accordion.mdx
@@ -1,19 +1,19 @@
-import { AccordionContainer, Accordion } from '@equisoft/design-elements-react';
+import { AccordionItem, Accordion } from '@equisoft/design-elements-react';
 import { ArgTypes, Meta, Stories, Story } from '@storybook/blocks';
 import * as AccordionStories  from './accordion.stories';
 
 
 # Accordion Component
 <Meta of={AccordionStories} />
-The Accordion component offers a robust Accessibility implementation, along with lots of flexibility with content and controls.<br />
-Each panel container has the capacity to accommodate a diverse range of content – spanning from basic text to intricate HTML elements.<br />
-With the flexibility to control whether multiple panels can remain open concurrently, or if the behavior restricts expansion to only one panel at a time. Also, control on enabling or disable specific accordion items.<br />
+The Accordion component offers a robust Accessibility implementation, along with lots of flexibility on content and controls.<br />
+Accordion controls whether multiple panels can remain open concurrently, or if the behavior restricts expansion to only one panel at a time.<br />
+Each AccordionItem has the capacity to accommodate a diverse range of content – spanning from basic text to intricate HTML elements.<br />
+Also, controls disable and collapse by default options implemented per item.<br />
 
 
-## Accordion Container Props
-All the props for the AccordionContainer can be utilized individually or in combination.
-<ArgTypes of={AccordionContainer} />
 ## Accordion Props
+<ArgTypes of={Accordion} />
+## Accordion Item Props
 <ArgTypes of={AccordionStories} />
 
 <Stories of={AccordionStories} />
